+++ Controllers/EmployeesController.cs
@@ -11,20 +11,32 @@ namespace App.Controllers
     public class EmployeesController : ControllerBase
     {
         private EmployeesService _service;
-        public EmployeesController(EmployeesService service)
+        private JwtService _jwtService;
+        public EmployeesController(EmployeesService service, JwtService jwtService)
         {
             _service = service;
+            _jwtService = jwtService;
         }
 
         [HttpGet]
-        public async Task<ActionResult<IList<EmployeeDTO>>> GetEmployees()
+        public async Task<ActionResult<IList<EmployeeDTO>>> GetEmployees([FromHeader] string Authorization)
         {
+            if (string.IsNullOrEmpty(Authorization) || !_jwtService.IsTokenValid(Authorization))
+            {
+                return Unauthorized();
+            }
+
             return Ok(await _service.FindAllEmployees());
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<EmployeeDTO>> GetEmployee(int id)
+        public async Task<ActionResult<EmployeeDTO>> GetEmployee(int id, [FromHeader] string Authorization)
         {
+            if (string.IsNullOrEmpty(Authorization) || !_jwtService.IsTokenValid(Authorization))
+            {
+                return Unauthorized();
+            }
+
             EmployeeDTO employee = await _service.FindEmployee(id);
             
             if (employee == null)
@@ -36,8 +48,13 @@ namespace App.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult<EmployeeDTO>> PostEmployee(Employee employee)
+        public async Task<ActionResult<EmployeeDTO>> PostEmployee(Employee employee, [FromHeader] string Authorization)
         {
+            if (string.IsNullOrEmpty(Authorization) || !_jwtService.IsTokenValid(Authorization))
+            {
+                return Unauthorized();
+            }
+
             EmployeeDTO postedEmployee = await _service.AddEmployee(employee);
 
             if (postedEmployee == null)
@@ -49,8 +66,13 @@ namespace App.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<ActionResult<EmployeeDTO>> PutEmployee(int id, Employee employee)
+        public async Task<ActionResult<EmployeeDTO>> PutEmployee(int id, Employee employee, [FromHeader] string Authorization)
         {
+            if (!string.IsNullOrEmpty(Authorization) && !_jwtService.IsTokenValid(Authorization))
+            {
+                return Unauthorized();
+            }
+
             if (id != employee.EmployeeNo)
             {
                 return BadRequest();
@@ -67,8 +89,13 @@ namespace App.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult<EmployeeDTO>> DeleteEmployee(int id)
+        public async Task<ActionResult<EmployeeDTO>> DeleteEmployee(int id, [FromHeader] string Authorization)
         {
+            if (string.IsNullOrEmpty(Authorization) || !_jwtService.IsTokenValid(Authorization))
+            {
+                return Unauthorized();
+            }
+
             EmployeeDTO deletedEmployee = await _service.RemoveEmployee(id);
 
             if (deletedEmployee == null)
